Moved the make_disc function to test_utils, might be handy for other tests.
